azure_functions/requirements.txt
# Azure Functions Core Dependencies
azure-functions>=1.18.0


# Azure Services
azure-ai-documentintelligence>=1.0.0
azure-core>=1.29.0
azure-identity>=1.15.0
azure-servicebus>=7.11.0
azure-storage-blob>=12.19.0
azure-cosmos>=4.5.1

# OpenAI
openai>=1.12.0

# Data Processing
tiktoken>=0.5.2

# HTTP Client
requests>=2.31.0

# Environment variables management
python-dotenv>=1.0.0

# JSON Web Token handling
PyJWT>=2.8.0
cryptography>=41.0.0

# Logging and monitoring
applicationinsights>=0.11.10
opencensus-ext-azure>=1.1.13
opencensus-ext-logging>=0.1.1

# Date and time utilities
python-dateutil>=2.8.2

# Type hints support
typing-extensions>=4.8.0

# JSON Schema validation
jsonschema>=4.20.0

# Path handling utilities
pathlib2>=2.3.7; python_version < '3.4'

# UUID generation
uuid>=1.30; python_version < '3.7'